Create and update for nested attributes

+++ db/migrate/20220930132216_create_trucks.rb
@@ -1,7 +1,7 @@
 class CreateTrucks < ActiveRecord::Migration[6.1]
   def change
     create_table :trucks do |t|
-      t.decimal :profit, default: 0
+      t.integer :profit, default: 0
 
       t.timestamps
     end
