+++ db/migrate/20221004042225_create_order_products.rb
@@ -4,7 +4,7 @@ class CreateOrderProducts < ActiveRecord::Migration[6.1]
       t.references :order, null: false, foreign_key: true
       t.references :product, null: false, foreign_key: true
       t.integer :quantity, default: 0
-      t.decimal :unit_price, default: 0
+      t.integer :unit_price, default: 0
 
       t.timestamps
     end
